Address code review feedback: remove unused deps and code, fix type casting

+++ app/src/main/java/com/netiface/nfsclient/NfsFetcher.kt
@@ -110,7 +110,7 @@ private class NfsBufferedSource(
                         -1L
                     } else {
                         buffer.write(data)
-                        position += data.size
+                        position += data.size.toLong()
                         data.size.toLong()
                     }
                 }
